게시물 삭제 기능 구현

+++ src/main/java/com/oyk/exam/jpaBoard/Controller/ArticleController.java
@@ -65,9 +65,24 @@ public class ArticleController {
     @RequestMapping("/doDelete")
     @ResponseBody
     public String doDelete(long id) {
+        if(articleRepository.existsById(id) == false) {
+            return """
+                    <script>
+                    alert('%d번 게시물은 이미 삭제되었거나 존재하지 않습니다.');
+                    history.back();
+                    </script>
+                    """.formatted(id);
+        }
+
         articleRepository.deleteById(id);
 
-        return "%d번 게시물이 삭제되었습니다.".formatted(id);
+        return """
+                    <script>
+                    alert('%d번 게시물이 삭제되었습니다.');
+                    location.replace('list');
+                    </script>
+                    """
+                .formatted(id);
     }
 
     @RequestMapping("/write")
